Move legacy xml code
