+++ tests/importexport/CMakeLists.txt
@@ -3,7 +3,7 @@ add_xtest(tst_datafileparser ${CMAKE_CURRENT_SOURCE_DIR}/csvdata.cpp)
 add_xtest_mock(tst_presethandler)
 add_xtest(tst_presetparser ${CMAKE_CURRENT_SOURCE_DIR}/presetfiletestdata.cpp)
 add_xtest(tst_projectfilejsonparser ${CMAKE_CURRENT_SOURCE_DIR}/projectfilejsontestdata.cpp)
-add_xtest(tst_projectfilexmlparser ${CMAKE_CURRENT_SOURCE_DIR}/projectfilexmltestdata.cpp)
+add_subdirectory(legacy)
 add_xtest(tst_projectfilejsonexporter)
 add_xtest(tst_projectfilehandler)
 add_xtest(tst_settingsauto ${CMAKE_CURRENT_SOURCE_DIR}/csvdata.cpp)
