+++ tests/importexport/legacy/CMakeLists.txt
@@ -1,0 +1,2 @@
+
+add_xtest(tst_projectfilexmlparser ${CMAKE_CURRENT_SOURCE_DIR}/projectfilexmltestdata.cpp)
